Refactor: move NumberNode to Number file for separating the NumberNode function from Client code
